schoolheat

+++ bbsapp/src/main/java/com/gaufoo/bbs/application/types/SchoolHeat.java
@@ -29,7 +29,7 @@ public interface SchoolHeat {
         PersonalInformation.PersonalInfo getLatestCommenter();
         Long getLatestActiveTime();
         Long getCreateTime();
-        Integer getHeat();
+        Long getHeat();
         Comment.AllComments getAllComments(Long skip, Long first);
     }
